i18n: dedupe useEyebrowClass, drop the deleteSemantics shim
The .eyebrow decision is needed by four areas (studio, library, training,
recording) but lived in the studio's panel primitives, and the recording area
had grown a second copy rather than import across that boundary. Moved to
hooks/useEyebrowClass and re-exported from primitives so the existing imports
keep working.

deleteSemantics' deprecated titlePrefix/description/confirmLabel fields are
gone now that SkillManageDialog — their only caller — uses the keys. The
English catalog import stays, type-only, so a renamed copy group is still a
compile error.
